Decompose Search.Sql emitters and lowering into cohesive units
SqlBuilder (1711 lines), Lower (930) and StructuralContext (548) each
carried most of a layer. Split them along the seams the project already
uses elsewhere, so all three lowering tiers now follow the same
dispatcher-plus-rules shape as Lowering/Leaf and Lowering/Composite.

- SqlBuilder -> 88 lines plus 7 emitters (predicate, plan validation,
  sort, CTE, include, shape, match page).
- Lower -> 475; include planning, resource-column extraction and the
  structural dispatch move out.
- StructuralContext -> 254; the CTE-graph accumulator separates from the
  rules that use it ($everything, chain, compartment, missing-param).
- Shared keyset-page rules collapse into KeysetPageInvariants as
  predicates only. Each layer keeps its own throw and wording: the two
  speak different vocabularies (options-level vs plan-level) and tests
  pin both messages. Sharing the rule kills drift where it is dangerous;
  tolerating message drift is harmless.
- The parallel SortKeyKind switches become a SortKeyEmitter strategy.
  SortEmitter keeps all phase and index policy.

Largest remaining file is 475 lines, down from 1711.

Emitted SQL is byte-identical, pinned by 1153 golden tests on two target
frameworks; no test file was touched. One deliberate exception, recorded
in the investigation doc: an undefined SortKeyKind now throws instead of
silently emitting a Date query, matching how undefined ChainDirection and
IncludeDirection values are already treated.

Copilot-Session: 705c14e2-ab7e-4fbc-870f-ba48a7cd5303
