+++ docs/features/search/readme.md
@@ -31,6 +31,7 @@ FHIR search is a complex domain requiring support for compartment searches, wild
 | [event-sourced-conformance](investigations/event-sourced-conformance.md) | Viable | Event-sourced architecture for package/conformance management - replaces 6 classes with 3, eliminates caches, atomic activation |
 | [includes-operation](investigations/includes-operation.md) | **Implemented** | `$includes` operation for paginated include/revinclude results with `_includesCount` parameter |
 | [not-referenced-search](investigations/not-referenced-search.md) | **Implemented** | `_not-referenced` parameter to find orphaned resources not referenced by others (SQL only) |
+| [search-sql-decomposition](investigations/search-sql-decomposition.md) | Viable | Structural decomposition of the three oversized `Ignixa.Search.Sql` stage orchestrators; removes duplicated plan invariants and extends the existing dispatcher/rule pattern upward |
 
 ## Related ADRs
